Refactor: rename progression.go to spawn.go, extract geom.go utilities
- Rename progression.go → spawn.go (content is structure placement/spawning, not general progression)
- Extract abs, chebyshevRingDo, and manhattan into new game/geom.go (geometry helpers shared across spawn.go and pathfinding.go)
- Move HasStructureOfType from State receiver in spawn.go to World receiver in world.go; keep thin State delegation in state.go so existing callers are unchanged
